Adding enums.go
